Add shape inference for v7 upsample; reuse the code
