+++ onnx/defs/tensor/defs.cc
@@ -1,8 +1,7 @@
 // Copyright (c) ONNX Project Contributors.
 // Licensed under the MIT license.
 
-#include "onnx/defs/schema.h"
-#include "onnx/defs/tensor_proto_util.h"
+#include "onnx/defs/tensor/utils.h"
 
 #include <algorithm>
 #include <cmath>
@@ -1324,69 +1323,6 @@ ONNX_OPERATOR_SET_SCHEMA(
           }
           return;
         }));
-
-void resizeShapeInference(InferenceContext& ctx) {
-  if (!hasNInputShapes(ctx, 1)) {
-    return;
-  }
-  propagateElemTypeFromInputToOutput(ctx, 0, 0);
-  auto& input_shape = getInputShape(ctx, 0);
-  auto* output_shape = getOutputShape(ctx, 0);
-  auto scales = ctx.getInputData(1);
-
-  if (output_shape->dim_size() > 0) {
-    if (output_shape->dim_size() != input_shape.dim_size()) {
-      fail_shape_inference(
-        "Ranks inferred (",
-        input_shape.dim_size(),
-        ") is not equal to the existing rank value (",
-        output_shape->dim_size(),
-        ").");
-    }
-  } else { // Infer the rank of output anyway
-    for (int i = 0; i < input_shape.dim_size(); ++i) {
-      output_shape->add_dim();
-    }
-  }
-  if (nullptr != scales) {
-    // Infer output shape's dimension value if 'scales' is known.
-    if (scales->data_type() == TensorProto::FLOAT) {
-      const auto& scales_data = ParseData<float>(scales);
-      if (scales_data.size() != static_cast<size_t>(input_shape.dim_size())) {
-        fail_shape_inference(
-          "Number of elements of input 'scales' must be same as rank of input 'X'");
-      }
-      for (int i = 0; i < input_shape.dim_size(); ++i) {
-        auto* dim = output_shape->mutable_dim(i);
-        // If input_shape has dim_value, we caculate the scaled result
-        // If input_shape doesn's have one, we leave it here
-        if (input_shape.dim(i).has_dim_value()) {
-          int64_t dim_value = static_cast<int64_t>(std::floor(
-            static_cast<float>(input_shape.dim(i).dim_value()) *
-            scales_data[i]));
-          // If output_shape has dim_value, we validate the caculated result
-          // If output_shape doesn's have one, we set it to the scaled result
-          if (dim->has_dim_value()) {
-            if (static_cast<int64_t>(dim->dim_value()) != dim_value) {
-              fail_shape_inference(
-                "Dimension value inferred (",
-                dim_value,
-                ") is not equal to the existing dim value (",
-                dim->dim_value(),
-                ").");
-            }
-          }
-          else {
-            dim->set_dim_value(static_cast<int64_t>(dim_value));
-          } // dim->has_dim_value()
-        } // input_shape.dim(i).has_dim_value()
-      } // empty_rank
-    } else {
-      fail_shape_inference(
-        "Input 'scales' must have float element type.");
-    } // nullptr != scales
-  }
-}
 
 static const char* Upsample_ver10_doc = R"DOC(
 Upsample the input tensor.
